Removed the bug that allowed users to pick up two or more miniatures at the same time
Hopefully removed the bug...the input manager now has a reference to whichever miniature is selected and the code now prevents the miniature button from creating a new miniature when the user is already holding one. The user should also no longer be able to pick up another miniature if another one is selected.